Speedup destruction of sheets with too many listeners & broadcasters.
Listeners and broadcasters are M:N relationship.  If you want to
destruct them, you easily end up in O(M*N) situation; where for every
listener, you iterate all broadcasters, to remove that one listener.

To avoid that, announce to the broadcasters that they are going to die, and the
listeners do not have to bother with removing themselves from the broadcaster.
The broadcaster will not broadcast anything after the PrepareForDesctruction()
call anyway.

Change-Id: I68d78b23e73bcbb944de9139448b2c20dfa14f62