Improve Accessibility mapping of roles for Mac
Since role Heading now is exposed to A11y we need to give it a better role then
unknown, otherwise it gets inaccessible to VoiceOver. Until I get a better grip
on how to best expose AXHeading, let's map it to textAreaRole so we at least
get back to the behaviour before IA2 integration.
Header and Footer is probably best mapped to AXGroup.

Change-Id: I6353f4d25c18e6e550df289ca2e999ad1c01f2a8
Reviewed-on: https://gerrit.libreoffice.org/8682
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>