sw: don't dispose SwAccessibles in SwRowFrm::Cut()
This does not appear to work so well: for a row with multiple cells,
when the second cell frame's SwAccessible is disposed it will send an
event, and the AtkListener will then request all children of the row,
and this causes a new SwAccessible to be created for the first cell
frame that was disposed previously.

Unlike most other events, Dispose events cannot be sent delayed.

The SwAccessibles are disposed from ~SwCellFrm already, that should be
sufficient.

Change-Id: I6da72605e5386540f8d82535fca151e97d70869f