discard more header/footer stuff when discarding headers/footers (bnc#875718)
782adba4d436c65cdf85d48f28486321873b15ed discards unneeded headers/footers,
but only the text, not e.g. frames or fields. The test document therefore
ends up with a frame with a page number inside the body text, because the text
in the footer it should be anchored to was discarded. This commit fixes
this by discarding more (although probably the whole header/footer
shouldn't even be parsed to begin with).

The test from b349d2483e1fe64316d87b55d0b3b4c8f2293e2e actually checked
for this incorrect frame, so change that (the whole test is suspicious, as
the end result is quite far away from what the original doc looks like).

(cherry picked from commit 5510f563502168defa4ccfc54214d781a7c92868)
Signed-off-by: Andras Timar <andras.timar@collabora.com>

Conflicts:
	sw/qa/extras/ooxmlexport/ooxmlexport.cxx
	sw/qa/extras/ooxmlimport/ooxmlimport.cxx

Change-Id: I2e7192e00237db1f030d0524c5667fe92c9e496b