Change SwFrmFmts to o3tl::sorted_vector like API
This changes the SwFrmFmts class std::vector inheritance to private
and extends the class to a o3tl::sorted_vector compatible API.

This should just be a cleanup patch and is a preparation for the
change of SwFrmFmts from vector to o3tl::sorted_vector.

For simple list cases, this also adds a SwFrmFmtsV, a std::vector
version of SwFrmFmts.

Change-Id: I2e91cb4d650b1c46c531885869d201edba84e5a6