SwDocShell: custom copy for embedded data source definition on save-as
If "EmbeddedDatabase" in test.odt refers test.ods in the same directory,
that will be "../../test.ods". Now if we save test.odt in a different
directory, we need to re-save the embedded data source definition,
otherwise the relative reference will resolve to a non-existing path.

Relative references are normally not supported for embedded objects, so
this is not a problem, but for data sources they are, that's why they
are a special case here.

Change-Id: Id138b9cdc38f2de589d9b80c66f1a61174699770
(cherry picked from commit 2d7ff7aabc1aa8cf5bb4900ae4b00feb8bc0f7f7)